refactor: move functions around, to split up huge _memoization_map module
